third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.limit.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify only one partitioned popin can be open at a time.
  assert_equals: expected "FirstPopinOpened-" but got "FirstPopinOpened-SecondPopinOpened-"
Harness: the test ran to completion.
